feat: CandleStickRepository에 CandleStickDBManager 생성
- dbManager를 통해 db에 저장되어 있는 특정 타입의 데이터를 반환하도록 요청
